requirements.txt
telethon
pydantic
python-dotenv
openai
httpx
jinja2
loguru
aiofiles
pytz
fastapi
uvicorn
streamlit
pandas
